requirements.txt
python-dotenv
pyrogram
tgcrypto
aiogram